Fix re-serialization in the export
Instead of passing false as the second param to unserialize, pass
an array where allowed classes are true, so that if the class does
exist, it will be instantiated.

Also check for __PHP_Incomplete_Class when re-serializing and ignore
those cases so that the execution doesn't error out and fail.